Mostly bug fixing & code cleanup
	* Added support for Moz API values to be passed via the command line
or by changing values in a new, separate config.php file.
	* Merged the two curlWrap functions into a single new
APICallWrapperClass
	* Renamed the SimpleBaseFileClass as SimpleScooterCSVFileClass
    * Fixed: If company name is not present in the file or the website
cannot be found, then a bunch of blank N/A rows appears in the results
(consultants)
    * Fixed: Blank rows in the CSV are actually loaded as record rows,
but then just fail later on.  Should be scrubbed up front.
	* Fixed: If inputfile contains a filename only, output files are
incorrectly to the root directory of the user's drive.
	* Fixed: Blank rows in the input CSV file cause bad and error records
to be produced in the output file.