Fixed column duplication issue for Crunchbase
Accidentally left a space in the column names for Crunchbase.  This
caused duplicate columns to appear in the results.

Changed returned columns back to being "<not set>" instead of null so
that the user can differentiate between an intentionally set empty
value and no result being returned.

Removed quantcast. from the monthly uniques column for matching
accuracy and consistency.