Rewrote getting multipage data from APIs to be non-recursive
-- Crunchbase was dying because it was recursing too many times.  New
version doesn't use recursion and therefore doesn't die for that
reason.
-- New version takes an array structure of the API call settings and
returns the data in that structure.  This helps keep the API from
changing a million times for all plugins if a new option is needed.
-- New getKeyURLString() function allows a child class to return a
string for it's user API key that the parent appends to the URL.
-- Has not been fully tested for all cases yet, so right now, only
Crunchbase uses the new fetchDatafromAPI calls.